# src/knowledge_base/api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Unified Knowledge Base System API
  description: |
    # Unified Knowledge Base System API
    
    This API provides access to the Unified Knowledge Base System, allowing you to:
    
    - Add, update, and retrieve documents
    - Query the knowledge base with natural language
    - Manage users and API keys
    - Monitor system health and performance
    - Perform administrative tasks
    
    ## Authentication
    
    Most endpoints require authentication using either:
    - API key (via `X-API-Key` header)
    - Bearer token (via `Authorization` header)
    
    ## WebSocket Support
    
    Real-time communication is available through WebSocket at `/ws/{client_id}`.
    
    ## Rate Limiting
    
    API requests are subject to rate limiting based on your account tier.
  version: 1.0.0
  termsOfService: https://example.com/terms/
  contact:
    name: API Support
    url: https://example.com/support
    email: support@example.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://api.example.com/v1
    description: Production server
  - url: https://staging-api.example.com/v1
    description: Staging server
  - url: http://localhost:8000
    description: Local development server
...
paths:
...
  /query:
    post:
      summary: Query the knowledge base
      description: Query the knowledge base with natural language
      operationId: queryKnowledgeBase
      tags:
        - Query
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryModel'
            examples:
              basicQuery:
                summary: Basic query
                value:
                  query: "What is the capital of France?"
                  top_k: 5
              streamingQuery:
                summary: Streaming query
                value:
                  query: "What is the capital of France?"
                  stream: true
      responses:
        '200':
          description: Query response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponseModel'
            text/event-stream:
              schema:
                type: string
                description: Streaming response for queries with stream=true
        '400':
          description: Invalid query format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /query/batch:
    post:
      summary: Batch query
      description: Perform multiple queries in a single request
      operationId: batchQuery
      tags:
        - Query
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                queries:
                  type: array
                  items:
                    $ref: '#/components/schemas/QueryModel'
              required:
                - queries
      responses:
        '200':
          description: Batch query ID
          content:
            application/json:
              schema:
                type: object
                properties:
                  batch_id:
                    type: string
        '400':
          description: Invalid request format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /query/batch/{batch_id}:
    get:
      summary: Get batch query results
      description: Get the results of a batch query
      operationId: getBatchResults
      tags:
        - Query
      parameters:
        - name: batch_id
          in: path
          description: Batch query ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Batch query results
          content:
            application/json:
              schema:
                type: object
                properties:
                  batch_id:
                    type: string
                  status:
                    type: string
                    enum: [pending, completed, failed]
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/QueryResponseModel'
                  error:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Batch query not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /query/similarity:
    post:
      summary: Similarity search
      description: Find chunks similar to the provided text
      operationId: similaritySearch
      tags:
        - Query
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimilaritySearchRequest'
      responses:
        '200':
          description: Similar chunks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChunkModel'
        '400':
          description: Invalid request format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /query/filtered:
    post:
      summary: Filtered query
      description: Query the knowledge base with advanced filtering
      operationId: filteredQuery
      tags:
        - Query
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilteredQueryRequest'
      responses:
        '200':
          description: Query response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponseModel'
        '400':
          description: Invalid request format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /query/suggest:
    get:
      summary: Query suggestions
      description: Get query suggestions based on a prefix
      operationId: suggestQueries
      tags:
        - Query
      parameters:
        - name: prefix
          in: query
          description: Query prefix to get suggestions for
          required: true
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of suggestions
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 5
      responses:
        '200':
          description: Query suggestions
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    QueryModel:
      type: object
      properties:
        query:
          type: string
          description: Query text
        filter:
          type: object
          description: Filter criteria
          additionalProperties: true
        top_k:
          type: integer
          description: Number of results to return
          minimum: 1
          maximum: 100
          default: 5
        stream:
          type: boolean
          description: Whether to stream the response
          default: false
      required:
        - query
    ChunkModel:
      type: object
      properties:
        id:
          type: string
          description: Chunk ID
        text:
          type: string
          description: Chunk text
        document_id:
          type: string
          description: Document ID
        metadata:
          type: object
          description: Chunk metadata
          additionalProperties: true
        score:
          type: number
          format: float
          description: Relevance score
      required:
        - id
        - text
        - document_id
    QueryResponseModel:
      type: object
      properties:
        query:
          type: string
          description: Original query
        answer:
          type: string
          description: Generated answer
        chunks:
          type: array
          items:
            $ref: '#/components/schemas/ChunkModel'
          description: Retrieved chunks
        metadata:
          type: object
          description: Response metadata
          additionalProperties: true
      required:
        - query
        - answer
        - chunks
    SimilaritySearchRequest:
      type: object
      properties:
        text:
          type: string
          description: Text to find similar chunks for
        filter:
          type: object
          description: Filter criteria
          additionalProperties: true
        top_k:
          type: integer
          description: Number of results to return
          minimum: 1
          maximum: 100
          default: 5
      required:
        - text
    FilteredQueryRequest:
      type: object
      properties:
        query:
          type: string
          description: Query text
        metadata_filters:
          type: object
          description: Metadata filters
          additionalProperties: true
        content_filters:
          type: array
          items:
            type: string
          description: Content filters (keywords)
        date_range:
          type: object
          properties:
            start:
              type: string
              format: date
              description: Start date
            end:
              type: string
              format: date
              description: End date
          description: Date range filter
        top_k:
          type: integer
          description: Number of results to return
          minimum: 1
          maximum: 100
          default: 5
      required:
        - query
        - metadata_filters
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
        error_type:
          type: string
          description: Error type
        status_code:
          type: integer
          description: HTTP status code
      required:
        - error
        - error_type
        - status_code
...
  responses:
    Unauthorized:
      description: Authentication failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Permission denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication
    
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token authentication
security:
  - ApiKeyAuth: []
  - BearerAuth: []